requirements.txt
cupy-cuda11x==11.4.0
fastrlock==0.8.1
numpy==1.24.0
